requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-cov==4.1.0
pyyaml==6.0.1